refactor: moved mouse grid functionality
- Created folder called 'mouse_grid' in the 'ui' folder
- Moved all mouse grid related files into 'mouse_grid' folder
- Split 'MouseGrid' class and 'MouseGridInputValidator' class into their own files
- The definition for 'MouseGridInputValidator' class now belongs in 'mouse_grid_input.py'

These changes are meant to improve readability for us. Any graphical UI changes occur in mouse_grid_ui.py. Input validation occurs in mouse_grid_input.py. Commands executed by mouse grid occur in mouse_grid_commands.py.
